moved landing components to landing components folder from root components folder and updated imports accordingly
