+++ app/(landing)/blog/components/PostPage.tsx
@@ -7,7 +7,7 @@ import PostHeader from '@/app/(landing)/blog/components/PostHeader';
 import SectionSeparator from '@/app/(landing)/blog/components/SectionSeparator';
 import PostTitle from '@/app/(landing)/blog/components/PostTitle';
 import { Post } from '@/lib/sanity.queries';
-import { LandingNavbar } from '@/components/landing-navbar';
+import { LandingNavbar } from '@/app/(landing)/components/landing-navbar';
 import { Footer } from '../../components/Footer';
 import styled from 'styled-components';
