+++ app/(landing)/components/landing-content.tsx
