+++ app/(landing)/components/landing-hero.tsx
