+++ app/(landing)/components/landing-quotes.tsx
