+++ app/(landing)/pricing/page.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { tier_data } from '@/app/data'; // Import your JSON data file
 import ZelosFeatureComparison from './ZelosFeatureComparison'; // Import your component
 import Image from "next/image"
-import { PricingSection } from '@/components/landing-pricing';
-import { LandingNavbar } from '@/components/landing-navbar';
+import { PricingSection } from './components/landing-pricing';
+import { LandingNavbar } from '@/app/(landing)/components/landing-navbar';
 import { Footer } from '../components/Footer';
 
 const Page: React.FC = () => {
